backend/src/utils/requirements.txt
opencv-python
numpy
tensorflow
Pillow
ultralytics
